Move InnerEnumerators into FlatArray.InnerFlatList
