+++ src/collections-flat-array/Collections.FlatArray/FlatArray.T.InnerFlatList/FlatArray.InnerFlatList.InnerEnumeratorEmpty.cs
@@ -1,0 +1,26 @@
+﻿namespace System.Collections.Generic;
+
+partial struct FlatArray<T>
+{
+    partial class InnerFlatList
+    {
+        private sealed class InnerEnumeratorEmpty : IEnumerator<T>
+        {
+            internal InnerEnumeratorEmpty() { }
+
+            public bool MoveNext() => false;
+
+            public T Current
+                =>
+                throw InnerExceptionFactory.EnumerationEitherNotStartedOrFinished();
+
+            object IEnumerator.Current
+                =>
+                throw InnerExceptionFactory.EnumerationEitherNotStartedOrFinished();
+
+            public void Reset() { }
+
+            public void Dispose() { }
+        }
+    }
+}
